sotoki: do not build link at load time, add StackExchangeDB class
Building links at import time:

- takes times
- takes space
- is prone to error

On SuperUser rendering a few pages is slower, since the query to retrieve
a related answer must scan more items.

StackExchangeDB is a class to replace a full ORM this also avoid to do queries
in the templates. This way if we change the schema it's easy to update the
queries.