Added mobile version of User profile
SE's responsiveness doesn't scale down to phone for User profile.
Instead they serve a different page layout using a different CSS.

Mobile stylesheet include a different top menu, full-with and custom layout for
questions list and User profile.
Users list is unchanged nor is the Tags list (actually it even lacks the topbar).
Question page is also adjusted slightly.

Except for the User profile with is broken without mobile CSS, we keep the regular,
responsive version on all pages ; but we add the topbar and menu.

Because mobile.css conflicts with regular CSS we have to dynamically remove it
and apply mobile one using JS. So we can't use mobile CSS on non-JS.
We reuse the Mathjax JS browser detection to switch between modes.

Because of the different layouts, we include both versions in HTML of the User profile
(which is small) and toggle which is displayed using CSS.

We should probably build a better, single layout responsive profile in the future

In addition, we slightly adapt margins on Posts for mobile