Using thread-specific Redis connections
Although Redis Client is thread safe, it has proven imptracticle to use a single
Redis Client and Pipeline over several threads.
Errors include disconnections and timeouts that a ConnectionPool can solve
but given Redis() expects to receive a matching number of results vs commands
and the fact that the pipeline is never locked, it fails to assert this and crash.

We're getting rid of all these by creating per-thread connection and pipeline that
are individually executed.

Now passing redis-url directly to Redis Client, thus enforcing the use of unix: scheme
for socket usage