Compress JSON data in redis using Snappy
JSON data in our redis store takes a good amount of our RAM usage.
Especially the questions excerpt and titles (tags ones as well).
Snappy allows us to gain some space here without sacrificing much performance.

We compress the whole JSON-encoded string that we store in redis.
To be able to uncompress it, we must store and retrieve it as bytes so we had
to drop redis-py auto decoding.

Additionally, we optimized the JSON a bit:
- we store User Id as int when possible
- we store tags list in questions as Tag ID ints instead of tag names
- we store CreationDate as timestamp instead of ISO formatted string