Removing questions from per-tag list that we won't display
In order to build per-tag list of questions, we have to store a list of all questions
using the tag in redis. this ends up using a lot of memory.

Eventually, we only use a substract of those list to build the paginated list of
questions for each tag. ATM, it's 100 pages of 15 questions so at most 1500 questions.
For SO's top tags list javascript, we are storing a list of 2M+ question IDs while
we'll only use 1500…

Unfortunately, we can't know until we're done parsing all questions which 1500 to keep
as we keep the most popular ones. So we have to store in redis anyway.

What this changes is that once this initial parsing is done, we drop all the non-1500
ones, freeing tons of RAM that will come in handy in future steps.

We're also now displaying the total number of questions in a tag page versus the
paginated max one.

We've also dropped the T:ID:{tag} keys on redis which were used for one-way mapping
of tag names to IDs. We actually need two way to use tag by Ids more effisciently
and this is both faster and more lightweight to store it in-python vs creating a lot
of keys.