Better Tag by ID handling
On SE, it is possible to link to Tag pages using their IDs.
It's not used muched: only 26 occurences in SO June 2021 but it exists.

We were creating redirects for Tag by Id to Tag by name in the Zim for that but for SO
it requires creating 61K redirects for 26 uses…

As we now have a Tag Name/ID matching in RAM, we are rewriting links to those ID links
to the proper Name ones instead.

Also handle the special case of a Tag declared unused (0 use) but which is actually used
Happens on SO 2021-06 with tag `imac`. Seems like the tag was first used during the DB
dump process, after Tags were dumped but before questions were.
In this rare occasion, we drop the tag on the question.