Replaced base linking with static relative links
Other tools such as zimcheck or Kiwix-JS have a hard time dealing with our <base />
based Zim because it's additional information they need to introduce and compare
every resource access with.

Although it was elegant, we decided to move away from the <base /> as it's easy to
switch this here compared to adapting other tools

This removes the <base /> node and now every href and src attributes are given the
`to_root` prefix that was before set in the base.
Rewriter is now informed about this prefix. It didn't required to before.