Moved users list to Python
In order to exclude users without interaction from being processed, we record active
user ids during the Posts First Pass.
This list of users_ids is then queried when looping over the complete list of users
from the XML file.

Additionnaly, we need a sorted list of Most Active Users (Highest scores) to build
the paginated list of users (up to 3,600 ATM).

We were previously relying on Redis for those usages:
- a set for the list of user IDS (`unsorted_users`)
- a sorted set scored by reputation (`sorted_users`) for the users list built after
user processing
- when testing the Activeness of a user, we'd query the unsorted set

This had a lot of impact:
- while adding to the unsorted set was performing OK, querying redis for activeness
added unnecessary syncrhonous network round-trips
- sorted sets seem to require a large amount of memory at insertion time on redis
- keeping a sorted list of 2.1M users (stackO) for the sole purpose of listing
the top 3,600 is overkill.

This changes it to a less elegant but more efficient solution:
- we store our unsorted user ids in a regular python set.
- tesing activeness is direct from the set.
- when parsing user XML, we build a caped list of the Top 3.6K users using a dict
- Once parsing is done, we cleanup the list of unsorted users