Experimenting lower sized queues
Until we can reliabily be informed of Item written in the zim, we have to accomodate
our queues to slow the process down and let the libzim catch up with what we feed it.

This sets an equal amount of threads (--threads / 2) for
- Processing (storing, querying and transforming data)
- Image handling (downloading-optimizing-uploading or downloading)
- Zim compression

Processing is non a --threads sized queue so about 2 slots per thread.
Images is on an unbounded queue
Zim compression is on a 10-sized queue at libzim level

libzim only uses a single thread for actually writting to Zim so our memory fills up
with binary images and HTML articles to add until those are actually written