Keep redis DB conditionnaly
Keeping redis database can be very useful for debugging or performance analysis
but obviously it shouldn't be the default.
This adds a --keep-redis option for that.

Another usage is for docker where we bundle redis and for which the process dies with
the scraper ; it can save a lot of time on SO to just keep it and let the OS reclaim
the RAM instead of manually flushing it