[TEST] using fixed number of workers
Increasing number of workers proved to be counter-productive as python spends
too much time trying to switch context rather than actually working.

Basic tests show that for the Questions step (65% of total), 3 workers is as fast as
working with slightly more workers. A lot of workers decreases speed.

This fixes nb workers to 3 and provides a static pool of image-related workers of 10
with a queue of 30.

Not alike other workers which do only text transformation from a single feed, image
workers do network operations that can stale for long.