Added retry on Redis Connection Error
It seems ConnectionError to redis (TCP or socket) are not that uncommon under certain
circumstances that havent been elucidated but probably involve threading and redis-py.

Online users seem to overcame this with a retry mechanism so introduces a retry
mechanism on our read-only calls.
ConnectionError always (twice) seemed to happen on `get_user_full()` during Questions Step
which requests a U: key to decode and json-decode it for addition in template.

This shall help us understand whether redis dies completely
or if it's just having a bad time.