TEST: increase image threads drastically
Latest SO run on pixelmemory showed the scraper was dragged down by the network
requests to download images.

This is an attempt to free the scrape back by starting 100 threads on a 200 queue instead of 10/30.
This would probably be more efficient if implemented as coroutines but this should be enough to
validate whether such an increase have a positive impact