TEST run garbage collector periodically
This runs the cycle-reference garbage collector (on gen2) periodically
during the two passes on the Posts data.
It is called every 10K articles viewed as it as an impact on performance:
it scans all the python variables several times to find out which refs
are not required anymore.

Need for this is unclear at the moment. This should help demonstrate
whether there is a need for a deeper investigation on cycling references

https://docs.python.org/3/library/gc.html#gc.collect