Faster images retrieval using single request
Instead of doing two requests to S3 to:
1. Verify if we have a matching key+metadata
2. Download the key's content

We are now directly requesting the content + metadata and comparing metadata upon reception.

- It has no impact on cases where we haven't uploaded the key.
- It has a negative impact on cases where we have an object in S3 for that key
but the metadata doesn't match -> updated image on same URL (probably not that frequent)
- It has a 50% positive impact on cases where we have the object with proper metadata.

So, to summarize:
- initial runs are not affected: images are not found on first request, created and uploaded.
- big speed improvement on sucessive runs as all non-modified images are fetched on a single request.
Those that have changed (on same URL) do waste some resources as those are downloaded before being discarded.
As most images are small, impact is limited.